TradeFragment Implementation UI Updated Candles charts Added

+++ app/src/main/java/ru/meseen/dev/stock/data/db/entitys/StockCandles.kt
@@ -3,19 +3,19 @@ package ru.meseen.dev.stock.data.db.entitys
 import ru.meseen.dev.stock.data.network.pojo.StockCandlesResponse
 
 data class StockCandles(
-    val close_prices: List<Double?>? = null,
+    val close_prices: List<Float?>? = null,
 
     val status_response: String? = null,
 
-    val timestamp: List<Int?>? = null,
+    val timestamp: List<Long?>? = null,
 
-    val volume_datas: List<Int?>? = null,
+    val volume_datas: List<Long?>? = null,
 
-    val high_prices: List<Double?>? = null,
+    val high_prices: List<Float?>? = null,
 
-    val low_prices: List<Double?>? = null,
+    val low_prices: List<Float?>? = null,
 
-    val open_prices: List<Double?>? = null
+    val open_prices: List<Float?>? = null
 ) {
     constructor(traStockCandles: StockCandlesResponse) : this(
         close_prices = traStockCandles.close_prices,
